config: extract helper function that loads revset/template aliases
I'm going to add fileset aliases, and we wouldn't want to maintain three
duplicated functions.

load_revset_aliases() is moved back to cli_util.rs because it's short and no
longer includes revset-specific logic.
